completed assignment per basic requirements
have an issue with masonry not working after the 1st search.  would
like to work on the aesthetics of the site.  a default search to
display,  style the return of the data, reg replace on the author
string to only display the author would be nice.  modal hasn’t been
implemented yet.